freetdm: multiple fixes for the core and analog 	* Replace ftdm_channel_flag_t with defines. We reached the 32bit limit where does not seem reliable           to keep using enum. 	* Flags member for ftdm_channel_t is now uint64_t 	* Added FTDM_CHANNEL_CALL_STARTED flag to indicate when the API user knows about a call 	* Refactored raw_data member for ftdm_sigmsg_t. If raw_data needs to be freed it must be requested explicitly           through the autofree member 	* Added collected member to ftdm_sigmsg_t for FTDM_SIGEVENT_COLLECTED data instead of using raw_data 	* Created define FTDM_DIGITS_LIMIT for DNIS/ANI digits limit 	* Fix some stat checks and outdated code in ftmod_analog 	* Refactored ftdm_channel_get_history_str API to return the time offsets and time since last state change 	* Do not send FTDM_SIGEVENT_STOP and FTDM_SIGEVENT_RELEASED on calls that were never reported to the user

+++ libs/freetdm/src/include/private/ftdm_core.h
@@ -411,7 +411,7 @@ struct ftdm_channel {
 	uint32_t extra_id;
 	ftdm_chan_type_t type;
 	ftdm_socket_t sockfd;
-	uint32_t flags;
+	uint64_t flags;
 	uint32_t pflags;
 	uint32_t sflags;
 	ftdm_alarm_flag_t alarm_flags;
